microservicio usuario

+++ TPE_Entrega1/Usuario/src/main/java/com/usuario/repository/UsuarioRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import com.usuario.model.Usuario;
 
-public interface UsuarioRepository extends RepoBase<Usuario, Integer>{
+public interface UsuarioRepository extends RepoBase<Usuario, Long>{
 	static List<String> sorteableFields = List.of("id_usuario", "nombre","apellido","nro_celular","email","roles");
